Add lump sums integration tests
