+++ test/files/SITIELM0001_AP_20210812105407541.dat
@@ -1,5 +1,5 @@
 B^2021-08-12^2^200^0001^SFIP^AP
-H^SFI00000001^01^SFIP000001^1^1000000001^GBP^100^RP00^GBP^SFIP^M12
-L^SFI00000001^100^2022^80001^DRD10^SIP00000000001^RP00^N^1^G00 - Gross value of claim^2022-12-01^2022-12-01^SOS273
-H^SFI00000002^03^SFIP000002^2^1000000002^GBP^100^RP00^GBP^SFIP^M12
-L^SFI00000002^100^2022^80001^DRD10^SIP00000000002^RP00^N^1^G00 - Gross value of claim^2022-12-01^2022-12-01^SOS273
+H^SFIP0000001^01^S000001^1^1000000001^GBP^100^RP00^GBP^SFIP^M12
+L^SFIP0000001^100^2022^80001^DRD10^SIP00000000001^RP00^N^1^G00 - Gross value of claim^2022-12-01^2022-12-01^SOS273
+H^SFIP0000002^03^S000002^2^1000000002^GBP^100^RP00^GBP^SFIP^M12
+L^SFIP0000002^100^2022^80001^DRD10^SIP00000000002^RP00^N^1^G00 - Gross value of claim^2022-12-01^2022-12-01^SOS273
