+++ test/files/SITIELM0001_AP_20210812105407542.dat
@@ -1,3 +1,3 @@
 B^2021-08-12^1^99^0001^SFIP^AP
-H^SFI00000001^01^SFIP000001^1^1000000001^GBP^100^RP00^GBP^SFIP^M12
-L^SFI00000001^100^2022^80001^DRD10^SIP00000000001^RP00^1^G00 - Gross value of claim^2022-12-01^2022-12-01^SOS273
+H^SFIP0000001^01^S000001^1^1000000001^GBP^100^RP00^GBP^SFIP^M12
+L^SFIP0000001^100^2022^80001^DRD10^SIP00000000001^RP00^1^G00 - Gross value of claim^2022-12-01^2022-12-01^SOS273
