+++ test/files/SITILSES_0001_AP_20220624212100913.dat
@@ -1,0 +1,5 @@
+B^2022-06-24^2^200^1^LSES^AP
+H^LSES0000001^001^L0000001^1000000001^1^100^RP00^GBP
+L^LSES0000001^100^2022^10501^EGF00^RP00^1^G00 - Gross value of claim^2022-12-01
+H^LSES0000002^001^L0000002^1000000002^1^100^RP00^GBP
+L^LSES0000002^100^2022^10501^EGF00^RP00^1^G00 - Gross value of claim^2022-12-01
