+++ test/files/SITILSES_0001_AP_20220624212100914.dat
@@ -1,0 +1,3 @@
+B^2022-06-24^1^99^1^LSES^AP
+H^LSES0000001^001^L0000001^1000000001^1^100^RP00^GBP
+L^LSES0000001^100^2022^10501^EGF00^RP00^1^G00 - Gross value of claim^2022-12-01
