+++ test/integration/blob-storage.test.js
@@ -2,8 +2,8 @@ describe('Blob storage tests', () => {
   let blobServiceClient
   let container
   const { BlobServiceClient } = require('@azure/storage-blob')
-  const blobStorage = require('../../../app/storage')
-  const blobStorageConfig = require('../../../app/config/storage')
+  const blobStorage = require('../../app/storage')
+  const blobStorageConfig = require('../../app/config/storage')
   const mockFileList = ['test1.dat', 'test2.dat']
   const testFileContents = 'This is a test file'
